+++ src/main/java/org/example/service/PostService.java
@@ -4,9 +4,12 @@ package org.example.service;
 import org.example.exception.NotFoundException;
 import org.example.model.Post;
 import org.example.repository.PostRepository;
+import org.example.repository.PostRepositoryImpl;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public class PostService {
   private final PostRepository repository;
